+++ sensors/src/commonMain/kotlin/regolith/sensors/location/LocationAccess.kt
@@ -1,0 +1,16 @@
+package regolith.sensors.location
+
+import kotlinx.coroutines.flow.Flow
+
+/**
+ * Provides access to the current device's location.
+ */
+interface LocationAccess {
+    /**
+     * Updates whenever the device's location changes.
+     *
+     * The frequency and distinctness of these updates is determined by the
+     * implementing class.
+     */
+    val locationUpdates: Flow<LocationState>
+}